# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão de Colaboradores",
        "description": "Documentação da API de Gestão de Colaboradores utilizando Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor API"
        }
    ],
    "paths": {
...
        "/api/employees": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Listar funcionários da empresa do usuário autenticado",
                "description": "Retorna uma lista de funcionários filtrada por parâmetros opcionais. O usuário autenticado só pode visualizar funcionários associados à sua empresa.",
                "operationId": "getEmployees",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrar por nome do funcionário (busca parcial)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "João"
                        }
                    },
                    {
                        "name": "position",
                        "in": "query",
                        "description": "Filtrar por cargo do funcionário",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Gerente"
                        }
                    },
                    {
                        "name": "hired_at",
                        "in": "query",
                        "description": "Filtrar por data de contratação (YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2023-01-01"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de funcionários retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Employee"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "description": "Total de registros",
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "per_page": {
                                                    "description": "Quantidade de registros por página",
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "current_page": {
                                                    "description": "Página atual",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "description": "Última página disponível",
                                                    "type": "integer",
                                                    "example": 5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Não autorizado (usuário não autenticado)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso proibido (usuário sem permissão)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Você não tem permissão para acessar estes dados."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro interno no servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao buscar os funcionários."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/{id}": {
            "get": {
                "tags": [
                    "Employees"
                ],
                "summary": "Exibir detalhes de um funcionário",
                "description": "Retorna os detalhes de um funcionário específico pertencente à empresa do usuário autenticado.",
                "operationId": "d53e117f7e819b6f063fc96ea910eb75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do funcionário",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do funcionário retornados com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/EmployeeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acesso não autorizado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Acesso não autorizado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Funcionário não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Funcionário não encontrado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/employees/import": {
            "post": {
                "tags": [
                    "Employees"
                ],
                "summary": "Importar colaboradores via arquivo CSV",
                "description": "Permite o upload de um arquivo CSV para importar colaboradores para a empresa do usuário autenticado.",
                "operationId": "0fb02484f6805567feb499534d4bbf18",
                "requestBody": {
                    "description": "Arquivo CSV contendo os colaboradores a serem importados.",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "O arquivo CSV para importação.",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Arquivo enviado e processamento iniciado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Arquivo enviado e processamento iniciado."
                                        },
                                        "status_id": {
                                            "description": "ID do status da importação.",
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro de validação no envio do arquivo.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Arquivo inválido ou não enviado."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EmployeeResource": {
                "title": "Employee Resource",
                "description": "Representação de um funcionário.",
                "properties": {
                    "id": {
                        "description": "ID do funcionário",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Nome do funcionário",
                        "type": "string",
                        "example": "João da Silva"
                    },
                    "position": {
                        "description": "Cargo do funcionário",
                        "type": "string",
                        "example": "Desenvolvedor"
                    },
                    "email": {
                        "description": "E-mail do funcionário",
                        "type": "string",
                        "example": "joao.silva@example.com"
                    },
                    "hired_at": {
                        "description": "Data de contratação do funcionário",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-15"
                    },
                    "company_id": {
                        "description": "ID da empresa do funcionário",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
            "Employee": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "company_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "example": "Jonh Doe"
                    },
                    "cpf": {
                        "type": "integer",
                        "example": "12345678901"
                    },
                    "email": {
                        "type": "string",
                        "example": "jonh.doe@email.com"
                    },
                    "position": {
                        "type": "string",
                        "example": "Desenvolvedor"
                    },
                    "hired_at": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-01-15"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-15T12:00:00Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-15T12:00:00Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insira o token no formato 'Bearer {seu_token}'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}